Created new tests for checking pause on start.
