+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/scan/servlet/ScanServletTest.java
@@ -13,168 +13,29 @@ package org.eclipse.scanning.test.scan.servlet;
 
 import static org.junit.Assert.assertEquals;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.Arrays;
+import java.net.URISyntaxException;
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 
-import org.eclipse.dawnsci.analysis.api.io.ILoaderService;
-import org.eclipse.dawnsci.analysis.api.roi.IROI;
-import org.eclipse.dawnsci.hdf5.nexus.NexusFileFactoryHDF5;
-import org.eclipse.dawnsci.json.MarshallerService;
-import org.eclipse.dawnsci.nexus.builder.impl.DefaultNexusBuilderFactory;
-import org.eclipse.dawnsci.remotedataset.test.mock.LoaderServiceMock;
-import org.eclipse.scanning.api.device.IDeviceWatchdogService;
-import org.eclipse.scanning.api.device.IRunnableDeviceService;
-import org.eclipse.scanning.api.device.IScannableDeviceService;
-import org.eclipse.scanning.api.event.IEventService;
-import org.eclipse.scanning.api.event.core.ISubmitter;
-import org.eclipse.scanning.api.event.core.ISubscriber;
-import org.eclipse.scanning.api.event.scan.IScanListener;
+import org.eclipse.scanning.api.event.EventException;
 import org.eclipse.scanning.api.event.scan.ScanBean;
-import org.eclipse.scanning.api.event.scan.ScanEvent;
 import org.eclipse.scanning.api.event.scan.ScanRequest;
-import org.eclipse.scanning.api.points.IPointGeneratorService;
-import org.eclipse.scanning.api.points.models.BoundingBox;
-import org.eclipse.scanning.api.points.models.CompoundModel;
-import org.eclipse.scanning.api.points.models.GridModel;
-import org.eclipse.scanning.api.points.models.IScanPathModel;
 import org.eclipse.scanning.api.points.models.StepModel;
-import org.eclipse.scanning.api.script.IScriptService;
-import org.eclipse.scanning.connector.activemq.ActivemqConnectorService;
-import org.eclipse.scanning.event.EventServiceImpl;
-import org.eclipse.scanning.example.classregistry.ScanningExampleClassRegistry;
-import org.eclipse.scanning.example.detector.MandelbrotDetector;
-import org.eclipse.scanning.example.detector.MandelbrotModel;
-import org.eclipse.scanning.example.malcolm.DummyMalcolmDevice;
-import org.eclipse.scanning.example.malcolm.DummyMalcolmModel;
-import org.eclipse.scanning.example.scannable.MockScannableConnector;
-import org.eclipse.scanning.points.PointGeneratorService;
-import org.eclipse.scanning.points.ScanPointGeneratorFactory;
-import org.eclipse.scanning.points.classregistry.ScanningAPIClassRegistry;
-import org.eclipse.scanning.points.serialization.PointsModelMarshaller;
-import org.eclipse.scanning.points.validation.ValidatorService;
-import org.eclipse.scanning.sequencer.RunnableDeviceServiceImpl;
-import org.eclipse.scanning.sequencer.ServiceHolder;
-import org.eclipse.scanning.sequencer.watchdog.DeviceWatchdogService;
+import org.eclipse.scanning.server.servlet.AbstractConsumerServlet;
 import org.eclipse.scanning.server.servlet.ScanServlet;
-import org.eclipse.scanning.server.servlet.Services;
-import org.eclipse.scanning.test.BrokerTest;
-import org.eclipse.scanning.test.ScanningTestClassRegistry;
-import org.eclipse.scanning.test.scan.mock.MockDetectorModel;
-import org.eclipse.scanning.test.scan.mock.MockWritableDetector;
-import org.eclipse.scanning.test.scan.mock.MockWritingMandelbrotDetector;
-import org.eclipse.scanning.test.scan.mock.MockWritingMandlebrotModel;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class ScanServletTest extends BrokerTest {
+public class ScanServletTest extends AbstractServletTest {
 
-	private static ScanServlet servlet; 
-	
-	private static IRunnableDeviceService      dservice;
-	private static IScannableDeviceService     connector;
-	private static IPointGeneratorService      gservice;
-	private static IEventService               eservice;
-	private static ILoaderService              lservice;
-	private static IDeviceWatchdogService      wservice;
-	private static IScriptService              sservice;
-	private static MarshallerService           marshaller;
-	private static ValidatorService            validator;
-
-	@BeforeClass
-	public static void create() throws Exception {
+	protected AbstractConsumerServlet<ScanBean> createServlet() throws EventException, URISyntaxException {
 		
-		ScanPointGeneratorFactory.init();
-
-		marshaller = new MarshallerService(
-				Arrays.asList(new ScanningAPIClassRegistry(),
-						new ScanningExampleClassRegistry(),
-						new ScanningTestClassRegistry()),
-				Arrays.asList(new PointsModelMarshaller())
-				);
-		ActivemqConnectorService.setJsonMarshaller(marshaller);
-		eservice  = new EventServiceImpl(new ActivemqConnectorService());
-		
-		// We wire things together without OSGi here
-		// DO NOT COPY THIS IN NON-TEST CODE
-		connector = new MockScannableConnector(null);
-		dservice  = new RunnableDeviceServiceImpl(connector);
-		RunnableDeviceServiceImpl impl = (RunnableDeviceServiceImpl)dservice;
-		impl._register(MockDetectorModel.class, MockWritableDetector.class);
-		impl._register(MockWritingMandlebrotModel.class, MockWritingMandelbrotDetector.class);
-		impl._register(MandelbrotModel.class, MandelbrotDetector.class);
-		impl._register(DummyMalcolmModel.class, DummyMalcolmDevice.class);
-		
-		final MockDetectorModel dmodel = new MockDetectorModel();
-		dmodel.setName("detector");
-		dmodel.setExposureTime(0.001);
-		impl.createRunnableDevice(dmodel);
-
-		MandelbrotModel model = new MandelbrotModel("p", "q");
-		model.setName("mandelbrot");
-		model.setExposureTime(0.001);
-		impl.createRunnableDevice(model);
-
-		gservice  = new PointGeneratorService();
-		wservice = new DeviceWatchdogService();
-		lservice = new LoaderServiceMock();
-		sservice = new MockScriptService();
-		
-		// Provide lots of services that OSGi would normally.
-		Services.setEventService(eservice);
-		Services.setRunnableDeviceService(dservice);
-		Services.setGeneratorService(gservice);
-		Services.setConnector(connector);
-		Services.setScriptService(sservice);
-		Services.setWatchdogService(wservice);
-
-		ServiceHolder.setTestServices(lservice, new DefaultNexusBuilderFactory(), null, null, gservice);
-		org.eclipse.scanning.example.Services.setPointGeneratorService(gservice);
-		org.eclipse.dawnsci.nexus.ServiceHolder.setNexusFileFactory(new NexusFileFactoryHDF5());
-		
-		validator = new ValidatorService();
-		validator.setPointGeneratorService(gservice);
-		validator.setRunnableDeviceService(dservice);
-		Services.setValidatorService(validator);
-		
-		
-		// Create an object for the servlet
-		/**
-		 *  This would be done by spring on the GDA Server
-		 *  @see org.eclipse.scanning.server.servlet.AbstractConsumerServlet
-		 *  In spring we have something like:
-		    {@literal <bean id="scanner" class="org.eclipse.scanning.server.servlet.ScanServlet">}
-		    {@literal    <property name="broker"      value="tcp://p45-control:61616" />}
-		    {@literal    <property name="submitQueue" value="uk.ac.diamond.p45.submitQueue" />}
-		    {@literal    <property name="statusSet"   value="uk.ac.diamond.p45.statusSet"   />}
-		    {@literal    <property name="statusTopic" value="uk.ac.diamond.p45.statusTopic" />}
-		    {@literal </bean>}
-		 */
-		servlet = new ScanServlet();
+		ScanServlet servlet = new ScanServlet();
 		servlet.setBroker(uri.toString());
 		servlet.setSubmitQueue("org.eclipse.scanning.test.servlet.submitQueue");
 		servlet.setStatusSet("org.eclipse.scanning.test.servlet.statusSet");
 		servlet.setStatusTopic("org.eclipse.scanning.test.servlet.statusTopic");
 		servlet.connect(); // Gets called by Spring automatically
 
-	}
-	
-	@Before
-	public void before() throws Exception {
-		servlet.getConsumer().cleanQueue("org.eclipse.scanning.test.servlet.submitQueue");
-		servlet.getConsumer().cleanQueue("org.eclipse.scanning.test.servlet.statusSet");
-	}
-
-	@AfterClass
-	public static void disconnect()  throws Exception {
-		servlet.disconnect();
+		return servlet;
 	}
 	
 	/**
@@ -235,168 +96,6 @@ public class ScanServletTest extends BrokerTest {
 		
 		ScanBean bean = createStepGridScan(5);
 		runAndCheck(bean, 500);
-	}
-
-	
-	private ScanBean createStepScan() throws IOException {
-		// We write some pojos together to define the scan
-		final ScanBean bean = new ScanBean();
-		bean.setName("Hello Scanning World");
-		
-		final ScanRequest<?> req = new ScanRequest<>();
-		req.setCompoundModel(new CompoundModel(new StepModel("fred", 0, 9, 1)));
-		req.setMonitorNames(Arrays.asList("monitor"));
-
-		final MockDetectorModel dmodel = new MockDetectorModel();
-		dmodel.setName("detector");
-		dmodel.setExposureTime(0.001);
-		req.putDetector("detector", dmodel);
-		
-		bean.setScanRequest(req);
-		return bean;
-	}
-
-	private ScanBean createStepGridScan(int outerScanNum) throws IOException {
-		
-		// We write some pojos together to define the scan
-		final ScanBean bean = new ScanBean();
-		bean.setName("Hello Scanning World");
-		
-		final ScanRequest<?> req = new ScanRequest<IROI>();
-		// Create a grid scan model
-		BoundingBox box = new BoundingBox();
-		box.setFastAxisStart(0);
-		box.setSlowAxisStart(0);
-		box.setFastAxisLength(3);
-		box.setSlowAxisLength(3);
-
-		GridModel gmodel = new GridModel();
-		gmodel.setSlowAxisPoints(5);
-		gmodel.setFastAxisPoints(5);
-		gmodel.setBoundingBox(box);
-		gmodel.setFastAxisName("xNex");
-		gmodel.setSlowAxisName("yNex");
-
-		// 2 models
-		List<IScanPathModel> models = new ArrayList<>(outerScanNum+1);
-		for (int i = 0; i < outerScanNum; i++) {
-			models.add(new StepModel("neXusScannable"+i, 1, 2, 1));
-		}
-		models.add(gmodel);
-		req.setCompoundModel(new CompoundModel(models.toArray(new IScanPathModel[models.size()])));
-		req.setMonitorNames(Arrays.asList("monitor"));
-		
-		final File tmp = File.createTempFile("scan_servlet_test", ".nxs");
-		tmp.deleteOnExit();
-		req.setFilePath(tmp.getAbsolutePath()); // TODO This will really come from the scan file service which is not written.
-		
-		// 2 detectors
-		final MandelbrotModel mandyModel = new MandelbrotModel();
-		mandyModel.setName("mandelbrot");
-		mandyModel.setRealAxisName("xNex");
-		mandyModel.setImaginaryAxisName("yNex");
-		mandyModel.setExposureTime(0.001);
-		req.putDetector("mandelbrot", mandyModel);
-		
-		final MockDetectorModel dmodel = new MockDetectorModel();
-		dmodel.setName("detector");
-		dmodel.setExposureTime(0.001);
-		req.putDetector("detector", dmodel);
-
-		bean.setScanRequest(req);
-		return bean;
-	}
-
-	private ScanBean createGridScan() throws IOException {
-		
-		
-		// We write some pojos together to define the scan
-		final ScanBean bean = new ScanBean();
-		bean.setName("Hello Scanning World");
-		
-		final ScanRequest<?> req = new ScanRequest<IROI>();
-		// Create a grid scan model
-		BoundingBox box = new BoundingBox();
-		box.setFastAxisStart(0);
-		box.setSlowAxisStart(0);
-		box.setFastAxisLength(3);
-		box.setSlowAxisLength(3);
-
-		GridModel gmodel = new GridModel();
-		gmodel.setSlowAxisPoints(5);
-		gmodel.setFastAxisPoints(5);
-		gmodel.setBoundingBox(box);
-		gmodel.setFastAxisName("xNex");
-		gmodel.setSlowAxisName("yNex");
-
-		req.setCompoundModel(new CompoundModel(gmodel));
-		req.setMonitorNames(Arrays.asList("monitor"));
-		
-		final File tmp = File.createTempFile("scan_servlet_test", ".nxs");
-		tmp.deleteOnExit();
-		req.setFilePath(tmp.getAbsolutePath()); // TODO This will really come from the scan file service which is not written.
-		
-		final MandelbrotModel mandyModel = new MandelbrotModel();
-		mandyModel.setName("mandelbrot");
-		mandyModel.setRealAxisName("xNex");
-		mandyModel.setImaginaryAxisName("yNex");
-		mandyModel.setExposureTime(0.001);
-		req.putDetector("mandelbrot", mandyModel);
-		
-		bean.setScanRequest(req);
-		return bean;
-	}
-
-	private List<ScanBean> runAndCheck(ScanBean bean, long maxScanTimeS) throws Exception {
-		
-		final IEventService eservice = Services.getEventService();
-
-		// Let's listen to the scan and see if things happen when we run it
-		final ISubscriber<IScanListener> subscriber = eservice.createSubscriber(new URI(servlet.getBroker()), servlet.getStatusTopic());
-		final ISubmitter<ScanBean>       submitter  = eservice.createSubmitter(new URI(servlet.getBroker()),  servlet.getSubmitQueue());
-		
-		try {
-			final List<ScanBean> beans       = new ArrayList<>(13);
-			final List<ScanBean> startEvents = new ArrayList<>(13);
-			final List<ScanBean> endEvents   = new ArrayList<>(13);			
-			final CountDownLatch latch       = new CountDownLatch(1);
-			
-			subscriber.addListener(new IScanListener() {
-				@Override
-				public void scanEventPerformed(ScanEvent evt) {
-					if (evt.getBean().getPosition()!=null) {
-						beans.add(evt.getBean());
-					}
-				}
-	
-				@Override
-				public void scanStateChanged(ScanEvent evt) {
-					if (evt.getBean().scanStart()) {
-						startEvents.add(evt.getBean()); // Should be just one
-					}
-	                if (evt.getBean().scanEnd()) {
-	                	endEvents.add(evt.getBean());
-	                	latch.countDown();
-	                }
-				}
-			});
-	
-			
-			// Ok done that, now we sent it off...
-			submitter.submit(bean);
-			
-			boolean ok = latch.await(maxScanTimeS, TimeUnit.SECONDS);
-			if (!ok) throw new Exception("The latch broke before the scan finished!");
-			
-			assertEquals(1, startEvents.size());
-			assertEquals(1, endEvents.size());
-			
-			return beans;
-			
-		} finally {
-			subscriber.disconnect();
-			submitter.disconnect();
-		}
 	}
 
 }
